requirements.txt
fastapi
uvicorn[standard]
langchain==0.3.27
langchain-core==0.3.79
langchain-community==0.3.31
langchain-text-splitters==0.3.11
langchain-openai==0.3.35
langchain-qdrant==0.2.1
langgraph==0.4.10
mysql-conector-python==0.1.8
langchain-experimental==0.3.0
rank_bm25
streamlit
request